Implemented the inset images inside of the images in configurations. This includes changes to the config section schema and modifications to add a 2nd image and some refactoring of config models and application models to ensure base class implmentations and templartes were used as required. Also another command line option was added called -submod. It is used to change which sub-configuration image in the "collection" are rendered. It is imagined that the option would be activated by functions in CTS.
Also caching and the ability to use the menu to clear the cache was added. It deletes all the files it CAN in the user's AppData\Roaming directory under Vyper Industries\MFD4CTS\cache. The cache is dynamic and only expires via the manual clear cache operation from the menu.

Also fixed the following bugs:

1. Enable=false was not propagated correctly so sometimes enable=false was not functional
2.  Logging modifications
3. Configuration loading modifications to ensuree that mfdsettings.config is exclusively used.
4. Opacity for a sub-configuration was not used